Fix types version issues, locks package versions, rename files and imports
