+++ package.json
@@ -38,8 +38,8 @@
     "react-dom": "~17.0.2",
     "react-ga4": "^2.1.0",
     "react-json-to-csv": "~1.0.4",
-    "react-leaflet": "^4.2.1",
-    "react-leaflet-draw": "^0.20.6",
+    "react-leaflet": "~3.2.5",
+    "react-leaflet-draw": "~0.20.0",
     "react-router-dom": "5.3.3",
     "react-scripts": "~5.0.1",
     "react-slick": "^0.30.3",
@@ -61,11 +61,11 @@
   },
   "devDependencies": {
     "@types/geojson": "^7946.0.14",
-    "@types/leaflet": "^1.9.16",
-    "@types/leaflet-draw": "^1.0.11",
+    "@types/leaflet": "~1.9.16",
+    "@types/leaflet-draw": "~1.0.11",
     "@types/node": "^17.0.41",
-    "@types/react": "^18.0.12",
-    "@types/react-dom": "^18.0.5",
+    "@types/react": "~17.0.2",
+    "@types/react-dom": "~17.0.2",
     "@types/react-json-to-csv": "^1.0.0",
     "@types/react-router-dom": "^5.3.3",
     "@types/styled-components": "^5.1.34",
