+++ src/declarations.d.ts
@@ -1,0 +1,2 @@
+// HACK: Mientras hacemos el cambio de carrusel a uno tipado
+declare module "react-slick";
