+++ src/pages/home/Carrousel.tsx
@@ -79,19 +79,25 @@ const modules: module[] = [
   },
   {
     id: 2,
+    title: "Monitoreo Comunitario",
+    image: Comunitario,
+    link: "/Monitoreo",
+    auth: false,
+  },
+  {
+    id: 3,
     title: "Indicadores de Biodiversidad",
     image: Indicadores,
     link: "/Indicadores",
     auth: false,
   },
   {
-    id: 3,
+    id: 4,
     title: "Portafolios",
     image: Portafolio,
     link: "/Portafolios",
     auth: false,
   },
-  //{ id: 4, title: "Monitoreo Comunitario", image: Comunitario, link: "/Monitoreo", auth: false },
   {
     id: 5,
     title: "Compensación Ambiental",
